relocate weather widget to correct folder
